Move cell-slab-iter to the legacy folder.
---
TYPE: IMPROVEMENT
DESC: Move cell-slab-iter to the legacy folder.
