+++ tiledb/sm/query/legacy/cell_slab_iter.h
